Replace locale model with locale element
