+++ lang/en.json
@@ -1,4 +1,5 @@
 {
     "Translate": "Translate",
-    "Enables multi-lingual websites.": "Enables multi-lingual websites."
+    "Enables multi-lingual websites.": "Enables multi-lingual websites.",
+    "Translate Page Properties": "Translate Page Properties"
 }
